feat: Render initial home page dynamically
The initial home page is no longer hard-coded in HTML. It is now being rendered dynamically using a separate function in the home.js file and imported into the main index.js file. The renderHome function in home.js is responsible for creating the header, main, and footer sections of the home page. The addTextContent and addInnerHTML functions are used to add text and HTML content to the elements.

Changes made:

Moved the home page content to the home.js file.
Created a renderHome function to dynamically render the content.
Created helper functions such as createElement, appendElement, addTextContent, and addInnerHTML to facilitate the rendering process.
Updated the index.js file to import renderHome function from home.js and called the renderHome function to render the initial home page content dynamically.